Consolidates setup_fogi methods into OpModel.setup_fogi.
After tweaking the implementations of the setup_fogi method in the various
model classes, this commit pulls together the large amount of common code
and creates a single OpModel.setup_fogi method which defers to derived
class _op_decomposition methods for the very small amount of derived-model-class
specific functionality.
